+++ docs/realm.js
@@ -151,6 +151,7 @@ class Realm {
     /**
      * Closes this Realm so it may be re-opened with a newer schema version.
      * All objects and collections from this Realm are no longer valid after calling this method.
+     * The method is idempotent.
      */
     close() { }